Filter directory tree during construction
I stumbled across this problem because I noticed that my Neuron directory
was super slow. `neuron new` took ~10s. I figured out that it was
slow because it was trying to read my `.direnv` into a `DirTree`.
`.direnv` contains a symlink to `nixpkgs`. Too many files!

Neuron was reading the entire directory into memory and applying filters
to that structure. When your directory tree has hundreds of thousands of
files, filtering the in-memory data structure is *too late*. Instead,
filter filters as the directory is being scanned.

This saves a lot of useless work, because large directories that should
be ignored are never read into memory.